+++ src/components/boards/risk/svg-region.vue
@@ -17,24 +17,4 @@
   stroke: #000000;
 }
 
-.p1 {
-   fill: #FF99C8;
-}
-
-.p2 {
-   fill: #FCF6BD;
-}
-
-.p3 {
-   fill: #D0F4DE;
-}
-
-.p4 {
-   fill: #A9DEF9;
-}
-
-.p5 {
-   fill: #E4C1F9;
-}
-
 </style>
