+++ src/components/chat/main.vue
@@ -1,7 +1,7 @@
 <template>
    <div class="chat">
       <messageThread :messages="messages"></messageThread>
-      <messageBox v-on:message="chatMessage"></messageBox>
+      <messageBox @message="chatMessage"></messageBox>
    </div>
 </template>
 
@@ -10,7 +10,12 @@ import messageThread from './messageThread'
 import messageBox from './messageBox'
 
 export default {
-   props: ['messages'],
+   props: ['socket', 'user'],
+   data : function () {
+      return {
+         messages: []
+      }
+   },
    name: 'Chat',
    template: "<Chat></Chat>",
    components: {
@@ -18,17 +23,30 @@ export default {
       messageBox
    },
    methods: {
+      // event handler for message chat
       chatMessage: function(msg) {
-         //console.log("EVENT" + msg)
-         this.$emit('message', msg)
+         let chatMessage = {
+            from: this.user,
+            msg: msg
+         }
+         this.messages.push(chatMessage)
+         this.socket.emit('message', chatMessage)
       }
+   },
+   mounted() {
+      this.socket.on('message', msg => {
+         // Incoming message added to current data.
+         this.messages.push(msg)
+      })
+
+
    }
 }
 </script>
 
 <style lang="css">
 
-.chat { margin: 0; padding: 0; box-sizing: border-box; }
+.chat { margin: 0; padding: 0; box-sizing: border-box; border:1px solid black;}
 .chat { font: 13px Helvetica, Arial; }
 
 </style>
